Individual functions
In this version, the check_and_fix function has been split into three separate functions: check_filenames, export_mismatches, and fix_mismatches. The check_filenames function is always run, and the export_mismatches and fix_mismatches functions are run if the corresponding flags are set. This makes the script more modular and easier to understand and maintain.